requirements.txt
APScheduler==3.3.0
Flask==0.11.1
gensim==0.12.4
nltk==3.2.1
numpy==1.11.2
pandas==0.19.1
pymongo==3.3.0
pytz==2016.4
Werkzeug==0.11.10
scikit_learn==0.18.1
python_twitter==3.2
